[FIX/#61] 이미지에서 데이터로
