+++ DoBday/DoBday/Model/Bday.swift
@@ -12,14 +12,14 @@ import SwiftData
 final class Bday: Identifiable {
     var id: UUID
     var name: String
-    var profileImage: String?
+//    var profileImage: String?
     var dateOfBday: Date?
     var isLunar: Bool
     var notiFrequency: [String]
     var relationshipTag: [String]
-//    @Attribute(.externalStorage) var profileImage: Data?
+    @Attribute(.externalStorage) var profileImage: Data?
 
-    init(id: UUID, name: String, profileImage: String? = nil, dateOfBday: Date? = nil, isLunar: Bool, notiFrequency: [String], relationshipTag: [String]) {
+    init(id: UUID, name: String, profileImage: Data?, dateOfBday: Date? = nil, isLunar: Bool, notiFrequency: [String], relationshipTag: [String]) {
         self.id = id
         self.name = name
         self.profileImage = profileImage
